chore: break apart WorkloadManager [DET-4716]
The new layer structure in the harness pipeline is:
 - SocketManager:    reads and writes to websocket (unchanged)
 - TimerLayer:       injects start_time/end_time into responses
 - TensorboardLayer: writes and syncs tensorboard metrics
 - StorageLayer:     prepares and uploads checkpoints
 - WorkloadManager:  sanity-checks and augments metrics responses

Previously, the last four responsibilities were all coded into the
WorkloadManager, even though they have little-to-nothing to do with each
other.

This also deletes the now-pointless Callback interface.
