+++ harness/determined/layers/__init__.py
@@ -1,5 +1,8 @@
 from determined.layers._harness_profiler import HarnessProfiler
 from determined.layers._socket_manager import SocketManager
+from determined.layers._storage import StorageLayer
+from determined.layers._tensorboard import TensorboardLayer
+from determined.layers._timer import TimerLayer
 from determined.layers._worker_process import (
     SubprocessLauncher,
     SubprocessReceiver,
