+++ harness/determined/load/__init__.py
@@ -8,5 +8,4 @@ from determined.load._load_trial_controller import (
     load_controller_from_trial,
     load_trial_implementation_controller,
     prepare_controller,
-    prepare_tensorboard,
 )
