+++ harness/determined/tensorboard/__init__.py
@@ -1,5 +1,5 @@
 from determined.tensorboard.base import TensorboardManager
 from determined.tensorboard.build import build, get_base_path, get_sync_path
-from determined.tensorboard.metric_writers import BatchMetricWriter, MetricWriter
+from determined.tensorboard.metric_writers import MetricWriter
 from determined.tensorboard.s3 import S3TensorboardManager
 from determined.tensorboard.shared import SharedFSTensorboardManager
